Fixed map_4mon_external_data.py
Number of issues, including:

Javascript parser doesn't support nan, so had to circumvent using number 
replacment in datafiles and when reading.

Properly handling releative path to jquary reading of json files.

Time format gets messed up in the writing and reading of the json files, 
so on read the 'date' values must be multiplied by 1e-6.